update(org-indent-bars-get-level-face): make leading stars invisible
In org-mode, it is possible to hide the leading stars of heading
lines by setting the variable `org-hide-leading-stars` to `t`.  What
happens is that the leading stars will have the face `org-hide` that
have in general the foreground color equal to the background color of
the face `default`.  This gives the effect to not seing them.

In our case, we want leading stars to be invisible because we already
use `line-prefix` text property to display xpm images with bars that
replace (visualy) those leading stars.

* `org-indent-bars-get-level-face`:  make leading stars invisible.
* `org-indent-bars-mode`:  add `org-indent-bars-invisible` to
  invisibility spec in `org-indent-bars-mode`.