update(org-indent-bars-get-level-face): modify star display with compose-region
Replacing the last leading star of each heading lines by modifying the
text property 'display works (visually).  But when you set this
property to nil to remove the visual replacement when desactivating
`org-indent-bars-mode`, it marks the buffer as modified, and we don't
want this.  This is really annoying.

But using `compose-region` and `decompose-region` functions we can
have the same visual result without marking the buffer as modified
when we desactivate the `org-indent-bars-mode`.  So we do.

* `org-indent-bars-remove-replacement-stars`:  New function.

* `org-indent-bars-stars`:  remove trailing space.
* `org-indent-bars-get-level-face`:  use `compose-region` function.
* `org-indent-bars-mode`:  update accordingly to the changes made in
  `org-indent-bars-get-level-face`.