update(tests): rename faces use in tests
When you want to completly remove a variable you can use `makunbound`,
when you want to completly remove a function you can use `fmakunbound`.
But in the case of a face, it seems not to be possible.

So it is better to name faces created in differents test with
different name.  This removes silly errors.