add(org-bars-set-line-properties): override its counterpart in org-indent-mode
In `org-indent-mode` the function `org-indent-set-line-properties`
using elements of the prefix arrays concatenate strings.  But in our
cases the element of prefix arrays aren't string but image data, so we
don't want to concatenate them.  We just want to add this image data
prefixes to the text properties `line-prefix` and `wrap-prefix`.